auth-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pyjwt==2.8.0
bcrypt==4.1.3